Working on viewModels

+++ app/src/main/java/uz/gita/budget_app/screens/main/report/monthly/intent_uistate/M_R_UiState.kt
@@ -4,9 +4,9 @@ import java.io.Serializable
 
 data class M_R_UiState(
     val monthData: String = "Loading",
-    val currentBalance: Long = 0,
+    val currentBalance: Float = 0f,
     val incomeCount: Long = 0,
     val expenseCount: Long = 0,
-    val previousBalance: Long = 0,
+    val previousBalance: Float = 0f,
     val monthlyReportList: List<Serializable> = emptyList()
 )
